CR: Extract snapshot clusterstate tests to their own class
